\name{amitriptyline}
\alias{amitriptyline}
\docType{data}
\title{Amitriptyline Data
}
\description{Patients who were admitted to the hospital after an amitriptyline overdose.}
\usage{data("amitriptyline")}
\format{
  A data frame with 17 observations on the following 7 variables.
  \describe{
    \item{\code{tot}}{Total TCAD plasma level.}
    \item{\code{ami}}{Amount of amitriptyline present in TCAD plasma level.}
    \item{\code{gen}}{Gender:1 if female, 0 if male.}
    \item{\code{amt}}{Amount of antidepressants taken at time of overdose.}
    \item{\code{pr}}{PR wave measurement.}
    \item{\code{diap}}{Diastolic blood pressure.}
    \item{\code{qrs}}{QRS wave measurement.}
  }
}
\details{
     This data set contains measures of side effects of amitriptyline and gender information.
     }
\references{
     Johnson, R.A., Wichern, D.W. (2007). Applied Multivariage Statistical Analysis, 6th edition.
     }

\keyword{datasets}
