\name{Berkeley}
\alias{Berkeley}
\docType{data}
\title{Berkeley Guidance Study Data
}
\description{Heights of children born in Berkeley}
\usage{data("Berkeley")}
\format{
  A data frame with 93 observations on the following 32 variables.
  \describe{
    \item{\code{V1}}{Sex.}
    \item{\code{V2}}{Age 1.}
    \item{\code{V3}}{Age 1.25.}
    \item{\code{V4}}{Age 1.5.}
    \item{\code{V5}}{Age 1.75.}
    \item{\code{V6}}{Age 2.}
    \item{\code{V7}}{Age 3.}
    \item{\code{V8}}{Age 4.}
    \item{\code{V9}}{Age 5.}
    \item{\code{V10}}{Age 6.}
    \item{\code{V11}}{Age 7.}
    \item{\code{V12}}{Age 8.}
    \item{\code{V13}}{Age 8.5.}
    \item{\code{V14}}{Age 9.}
    \item{\code{V15}}{Age 9.5.}
    \item{\code{V16}}{Age 10.}
    \item{\code{V17}}{Age 10.5.}
    \item{\code{V18}}{Age 11.}
    \item{\code{V19}}{Age 11.5.}
    \item{\code{V20}}{Age 12.}
    \item{\code{V21}}{Age 12.5.}
    \item{\code{V22}}{Age 13.}
    \item{\code{V23}}{Age 13.5.}
    \item{\code{V24}}{Age 14.}
    \item{\code{V25}}{Age 14.5.}
    \item{\code{V26}}{Age 15.}
    \item{\code{V27}}{Age 15.5.}
    \item{\code{V28}}{Age 16.}
    \item{\code{V29}}{Age 16.5.}
    \item{\code{V30}}{Age 17.}
    \item{\code{V31}}{Age 17.5.}
    \item{\code{V32}}{Age 18.}
  }
}
\details{
     This data set contains measurements of heights of children born in 1928-29 in Berkeley, CA. 
     }
\references{
	Tuddenham, R. D. and Snyder, M. M. (1954). Physical growth of California boys and girls from birth to eighteen years. Publications in child developments. University of California, Berkeley, 1(2), 183-364.
     }

\keyword{datasets}
