\name{horseshoecrab}
\alias{horseshoecrab}
\docType{data}
\title{Horseshoe Crab Data
}
\description{The number of satellite male crabs near a female crab upon characteristic of the female horseshoe crabs. }
\usage{data("horseshoecrab")}
\format{
  A data frame with 173 observations on the following 5 variables.
  \describe{
    \item{\code{V1}}{Color.}
    \item{\code{V2}}{Condition of spine.}
    \item{\code{V3}}{Width of shell.}
    \item{\code{V4}}{Satellite.}
    \item{\code{V5}}{Weight.}
  }
}
\details{
     This data set contains the number of satellite male crabs and characteristics of the female horseshoe crabs.
     }
\references{
     Agresti, A. (2007). An Introduction to Categorical Data Analysis, 2nd edition.
     }

\keyword{datasets}
