\name{sales}
\alias{sales}
\docType{data}
\title{Sales staff Data
}
\description{On the performance of a firm's sales staff }
\usage{data("sales")}
\format{
  A data frame with 50 observations on the following 7 variables.
  \describe{
    \item{\code{V1}}{Index of sales growth.}
    \item{\code{V2}}{Index of sales profitability.}
    \item{\code{V3}}{Index of new account sales.}
    \item{\code{V4}}{Score on creativity.}
    \item{\code{V5}}{Score on mechanical reasoning test.}
    \item{\code{V6}}{Score on abstract reasoning test.}
    \item{\code{V7}}{Score on Mathematics test.}
  }
}
\details{
     This data set contains 3 measures of performance and 4 tests scores.
     }
\references{
     Johnson, R.A., Wichern, D.W. (2007). Applied Multivariage Statistical Analysis, 6th edition.
     }

\keyword{datasets}
