\name{waterstrider}
\alias{waterstrider}
\docType{data}
\title{Water strider data
}
\description{Measures of characteristics of the water striders }
\usage{data("waterstrider")}
\format{
  A data frame with 90 observations on the following 9 variables.
  \describe{
    \item{\code{V1}}{Index of water strider species.}
    \item{\code{V2}}{Logarithm of length of the first antennal segment.}
    \item{\code{V3}}{Logarithm of length of the second antennal segment.}
    \item{\code{V4}}{Logarithm of length of the third antennal segment.}
    \item{\code{V5}}{Logarithm of length of the fourth antennal segment.}
    \item{\code{V6}}{Logarithm of length of fomora of middle leg.}
    \item{\code{V7}}{Logarithm of length of tibiae of middle leg.}
    \item{\code{V8}}{Logarithm of length of fomora of hind leg.}
    \item{\code{V9}}{Logarithm of length of tibiae of hind leg.}
  }
}
\details{
     This data set contains 8 measures of water striders and an indicator of the species of water striders.
     }
\references{
     Klingenberg, C. R. and Spence, J. R. (1993). Heterochrony and Allometry Lessons from the Water Strider Genus Limnoporus. Evolution 47, 1834-1853
     }

\keyword{datasets}
